Lock node version to match targeted lambda env
Ran across this lock with checking an environment variable for
undefined.  In whatever node version my system has set as default, an
environment variable that is undefined is set to the string undefined
breaking a check.  This should be revisited later, but for now, let's
just lock the node version.